👌 [RUM-232] move worker "boot" logic in a dedicated module
This follow other packages design
